Migrated to AWS CLI tools from ec2-api-tools and eliminated credentials from config
It is far more convenient to install aws cli (pip install awscli) and a lot better to configure profiles in a standard matter (~/.aws/config).
With this approach the user that does backups can also access the tools manually in the console and perform checks or other tasks.
For reference on how aws cli should be configured have a look here: https://github.com/aws/aws-cli